Add a notice in the title if an issue has been closed
A bit of a workaround, since detecting if an issue has been
closed/merged/reopened is a bit of a workaround.

Also, as far as I can see, there is no way to close an issue from the
GitHub API. But I'll ask around and see if it is possible.