Remove parantheses from `if` statements (and other minor fixes)
When in Python, do as the Pythonians do.

I left parantheses in for some of the more "complex" if statements just
because I feel it makes them easier to understand.

Also remove a `print()` left in from debugging, tweaked one comment's
grammar, and changed a `!= None` to an `is not None`.